requirements.txt
analytics-python==1.1.0
Flask==0.10.1
Flask-Login==0.3.0
flask-mongoengine==0.7.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mongoengine==0.10.0
newrelic==2.56.0.42
pymongo==3.0.3
python-dateutil==2.4.2
requests==2.8.1
sendgrid==1.5.3
six==1.9.0
smtpapi==0.2.0
SQLAlchemy==1.0.8
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
